Moved zipcodes to own app
