+++ .gitignore
@@ -9,6 +9,7 @@ userAuth/migrations
 houseDatabase/migrations
 homePage/migrations
 survey/migrations
+zipcodes/migrations
 
 # Visual Studio Code Settings #
 .vscode/
