+++ Cocoon/settings/base.py
@@ -35,6 +35,7 @@ INSTALLED_APPS = [
     'userAuth',
     'survey',
     'houseDatabase',
+    'zipcodes',
 ]
 
 MIDDLEWARE = [
